refactor(gen_reply): split one reply turn out of the cog
The cog kept only what belongs to the process; each phase of a turn now
owns its own module and class.
